// packages/client/native.d.ts
/* tslint:disable */
/* eslint-disable */

/* auto-generated by NAPI-RS */

/** A ':' separated key value pair */
export interface KeyValue {
  key: string
  value: string
}
export interface ClientOptions {
  /**
   * Additional headers.
   *
   * Values should be key value pairs separated by ':'
   */
  headers: Array<KeyValue>
  /** Username */
  username?: string
  /** Password */
  password?: string
  /** Auth token. */
  token?: string
  /** Server host. */
  host: string
}
export declare function createFlightSqlClient(
  options: ClientOptions
): Promise<FlightSqlClient>
export declare function rustCrateVersion(): string
export declare class FlightSqlClient {
  query(
    query: string,
    params?: Array<[string, string]> | undefined | null
  ): Promise<Buffer>
}